+++ tests/settings.py
@@ -1,4 +1,6 @@
 import os
+import random
+import string
 
 import django
 
@@ -6,7 +8,8 @@ BASE_DIR = os.path.dirname(os.path.dirname(__file__))
 DEBUG = False
 TEMPLATE_DEBUG = DEBUG
 
-SECRET_KEY = "TH1Sr3allyISn0tAsecretKETbut1SonlyUSEDforTeSTING"
+characters = string.ascii_letters + string.digits
+SECRET_KEY = "".join(random.choice(characters) for i in range(48))
 
 DATABASES = {
     "default": {